Simplify target resolution & release tasks 0.4.6
MDProfileVersion is no longer needed, it will be determined
from the target instance. If provided, it will be validated.

MDAppDir is a new convenience property for Mac that will
set MDBinDir from the .app path.
