+++ MonoDevelop.Addins.Tasks/MonoDevelop.Addins.Tasks.csproj
@@ -4,7 +4,7 @@
     <AppendTargetFrameworkToOutputPath>False</AppendTargetFrameworkToOutputPath>
     <Title>MonoDevelop Extension Build Support</Title>
     <PackageId>MonoDevelop.Addins</PackageId>
-    <PackageVersion>0.4.5</PackageVersion>
+    <PackageVersion>0.4.6</PackageVersion>
     <Authors>Mikayla Hutchinson</Authors>
     <PackageLicenseUrl>http://opensource.org/licenses/MIT</PackageLicenseUrl>
     <PackageProjectUrl>https://github.com/mhutch/MonoDevelop.AddinMaker</PackageProjectUrl>
